Handle Config.uiscale as a double, for consistency's sake.

+++ src/haven/Config.java
@@ -52,7 +52,7 @@ public class Config {
     public static int mainport = getint("haven.mainport", 1870);
     public static int authport = getint("haven.authport", 1871);
     public static boolean softres = getprop("haven.softres", "on").equals("on");
-    public static Float uiscale = getfloat("haven.uiscale", null);
+    public static Double uiscale = getfloat("haven.uiscale", null);
     public static byte[] authck = null;
     public static String prefspec = "hafen";
     public static final String confid = "";
@@ -81,11 +81,11 @@ public class Config {
 	}
     }
 
-    private static Float getfloat(String name, Float def) {
+    private static Double getfloat(String name, Double def) {
 	String val = getprop(name, null);
 	if(val == null)
 	    return(def);
-	return(Float.parseFloat(val));
+	return(Double.parseDouble(val));
     }
 
     private static void usage(PrintStream out) {
